+++ components/faphub/dao/api/src/main/java/com/flipperdevices/faphub/dao/api/FapReportApi.kt
@@ -1,0 +1,5 @@
+package com.flipperdevices.faphub.dao.api
+
+interface FapReportApi {
+    suspend fun report(applicationUid: String, description: String)
+}