+++ components/faphub/dao/network/src/main/java/com/flipperdevices/faphub/dao/network/retrofit/utils/FapHubNetworkCategoryApi.kt
@@ -35,7 +35,11 @@ class FapHubNetworkCategoryApi(
         }
         val categoriesReceived = try {
             categoryApi.getAll(
-                sdkApi = target.getApiForServer()
+                sdkApi = when (target) {
+                    is FlipperTarget.Received -> target.sdk.toString()
+                    FlipperTarget.Unsupported,
+                    FlipperTarget.NotConnected -> null
+                }
             )
         } catch (ex: Exception) {
             error(ex) { "Failed get categories" }