+++ components/faphub/fapscreen/impl/src/main/java/com/flipperdevices/faphub/fapscreen/impl/api/FapScreenApiImpl.kt
@@ -59,6 +59,7 @@ class FapScreenApiImpl @Inject constructor(
             deepLinks = deeplinkArguments
         ) {
             ComposableFapScreen(
+                navController = navController,
                 onBack = navController::popBackStack,
                 installationButton = { fapItem, modifier ->
                     installationUIApi.ComposableButton(