+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - [Feature] Deeplik on open fap application
 - [Feature] Add url to release catalog
 - [Feature] Add build status card to app screen
+- [Feature] Add report app button
 - [Refactor] Migrate all feature components modules from KSP to Anvil
 - [Refactor] Migrate to Ktorfit
 - [Refactor] Migrate key emulate to new module, decompose Emulate Helper